rename root components to index.tsx
