+++ client/src/App.tsx
@@ -4,7 +4,7 @@ import Container from '@material-ui/core/Container';
 import { makeStyles } from '@material-ui/core/styles';
 import EventListPage from './components/event-list-page/EventListPage';
 import EventPage from './components/event-page/EventPage';
-import AdminPage from './components/admin-page/AdminPage';
+import AdminPage from './components/AdminPage';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 const useStyles = makeStyles({
